backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0
numpy==1.26.4
Pillow==10.3.0
python-multipart==0.0.9
